+++ previous-versions/assess/src/codatassess/models/operations/get_commerce_customer_retention_metrics.py
@@ -3,9 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import periodunit as shared_periodunit
-from ..shared import report as shared_report
+from ...models.shared import periodunit as shared_periodunit
+from ...models.shared import report as shared_report
 from typing import Optional
 
 
@@ -33,12 +32,10 @@ class GetCommerceCustomerRetentionMetricsRequest:
 class GetCommerceCustomerRetentionMetricsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     report: Optional[shared_report.Report] = dataclasses.field(default=None)
     r"""OK"""