+++ previous-versions/assess/src/codatassess/models/operations/list_data_type_data_integrity_details.py
@@ -3,9 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import dataintegritydatatype as shared_dataintegritydatatype
-from ..shared import details as shared_details
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import dataintegritydatatype as shared_dataintegritydatatype
+from ...models.shared import details as shared_details
 from typing import Optional
 
 
@@ -31,13 +30,11 @@ class ListDataTypeDataIntegrityDetailsRequest:
 class ListDataTypeDataIntegrityDetailsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     details: Optional[shared_details.Details] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""